requirements.txt
PyYAML
PyMySQL
dearpygui
dearpygui_ext